+++ sandbox/ssr/src/routeMiddleware.ts
@@ -7,11 +7,14 @@ declare module "kiru/router" {
   }
 }
 
-export const requireAuth: RouteMiddleware = ({ context, meta }) => {
-  if (!meta.requiresAuth) return
+export const requireAuth: RouteMiddleware = ({ context, to }) => {
+  if (!to.meta.requiresAuth) return
   const user = context.user
   if (user) return
-  const login = typeof meta.unauthorizedRedirect === "string" ? meta.unauthorizedRedirect : "/login"
+  const login =
+    typeof to.meta.unauthorizedRedirect === "string"
+      ? to.meta.unauthorizedRedirect
+      : "/login"
   return { redirect: login }
 }